Add explanation that Docker port is 80 (#1480)
When compiling via `yarn`, the default port used is 3000 (see #L50).  However, the given Docker command runs the UI on port 80 (default HTTP port).  

I added an extra sentence enumerating this difference, because I used the same port with Docker as I did with the local build at first, and it was a bit confusing why I was getting a 404.